moved things around a bit, broken refactor, needs fixing
